1. Use fanVector<float,2> instead of fanVector<float,3> for fanLens::project 2. Add fanLens::getTransformation 3. Add constructor in fanVector & fanMatrix for construct from vector/matrix with different T and DIMENS 4. SDL use top left as origin point, fanFilm use bottom left, so translate it.

+++ src/camera/PointScannerCamera.cpp
@@ -5,7 +5,7 @@ using namespace fan;
 void PointScannerCamera::takePicture( fanScene& scene,
                                       fanFilm& film,
                                       fanLens& lens ) {
-    fanVector<int, 2> pos;
+    fanVector<float, 2> pos;
     for ( auto itor = scene.mVertices.begin(), end = scene.mVertices.end();
             itor != end; ++itor ) {
         if ( !lens.project( pos, *itor, film ) ) {
